mocked requests; created test_online_resources
I removed the old test for get_nwis() because it actually called the
nwis. I put it into test_online_resources.py and gave it a name that
won't get automatically run by nose2.  I also used a mock to repace the
call to requests, and tested that the request was made properly.