fixed test_online_resources.py to allow for use of pytest
I made it so that pytest will not discover my test of online resources.
Use this file to check whether an online service is up and running. Do
not check these outside resources during unit tests- it is irrelevant
and will slow down the unittests and possibly annoy the service
providers!