alter return signature of extract_nwis_df
hf.extract_nwis_df(json) now returns a dataframe and a dict with metadata.

This breaks so many tests! And likely most of the help docs.

Although most of the information in the dict can be gotten from the json easily without mixing this into the dataframe construction, you can only get the local freq when the dataframe is being constructed. So I mix the two jobs into one function.